review tweaks

+++ code/ship/ship.h
@@ -1126,7 +1126,7 @@ public:
 	char			generic_debris_pof_file[MAX_FILENAME_LEN]; // smaller debris bits thrown around willy-nilly on death
 	int				generic_debris_model_num;
 	int				generic_debris_num_submodels;
-	float			generic_debris_spew_num;
+	int			    generic_debris_spew_num;
 
 	// subsystem information
 	int		n_subsystems;						// this number comes from ships.tbl
